@@ -1,50 +1,87 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Eco.EM.Framework</RootNamespace>
-    <Version>3.3.2</Version>
-    <PackageId>ElixrMods.Framework</PackageId>
-    <Authors>Elixr Solutions</Authors>
-    <Product>EM Framework</Product>
-    <Copyright>Elixr Solutions 2020</Copyright>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://elixrmods.com/</PackageProjectUrl>
-    <RepositoryUrl>https://git.elixr-solutions.com/elixrmods/elixr-mods</RepositoryUrl>
-    <RepositoryType>Private Gitlab</RepositoryType>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <AssemblyName>Eco.EM.Framework</AssemblyName>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<RootNamespace>Eco.EM.Framework</RootNamespace>
+		<Version>3.4.2</Version>
+		<PackageId>ElixrMods.Framework</PackageId>
+		<Authors>Elixr Solutions</Authors>
+		<Product>EM Framework</Product>
+		<Copyright>Elixr Solutions 2020</Copyright>
+		<PackageLicenseExpression>MIT</PackageLicenseExpression>
+		<PackageProjectUrl>https://elixrmods.com/</PackageProjectUrl>
+		<RepositoryUrl>https://github.com/TheKye/EM-Framework/</RepositoryUrl>
+		<RepositoryType>Github</RepositoryType>
+		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+		<PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+		<AssemblyName>Eco.EM.Framework</AssemblyName>
+	</PropertyGroup>
 
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <OutputPath>..\bin\9.6\Framework\</OutputPath>
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputPath>..\bin\9.6\Framework\</OutputPath>
+		<Description>The EM Framework is a bundle of Tools and utilities for Modders of the Game Eco in order to improve their modding experience</Description>
+		<PackageIcon>EMICON-text.png</PackageIcon>
+		<PackageReadmeFile>README.md</PackageReadmeFile>
+		<PackageReleaseNotes>
+			-Added New Writing Utils
+			-Moved to HttpClient from WebRequest
+			-Best Practices Cleanup
+			-Minor Fixes
+			-Added Minor Migration for Housing Resolver
+			-Fixed Non static recipes being marked as static due to incorrect parameter
+			-Fixed LuckBreakGlobal not applying LuckyBreak
+			-Changed EcoPediaGenerator to use new writing utils
+			-Version Change
+		</PackageReleaseNotes>
+	</PropertyGroup>
+	<Target Name="PostBuild" AfterTargets="PostBuildEvent" Condition="'$(OS)' == 'Windows_NT'">
+		<Exec Command="cd $(TargetDir)&#xD;&#xA;del *.config&#xD;&#xA;del *.json&#xD;&#xA;del A*&#xD;&#xA;del B*&#xD;&#xA;del C*&#xD;&#xA;del D*&#xD;&#xA;del F*&#xD;&#xA;del G*&#xD;&#xA;del H*&#xD;&#xA;del I*&#xD;&#xA;del J*&#xD;&#xA;del K*&#xD;&#xA;del L*&#xD;&#xA;del M*&#xD;&#xA;del N*&#xD;&#xA;del O*&#xD;&#xA;del P*&#xD;&#xA;del Q*&#xD;&#xA;del R*&#xD;&#xA;del S*&#xD;&#xA;del T*&#xD;&#xA;del U*&#xD;&#xA;del V*&#xD;&#xA;del W*&#xD;&#xA;del X*&#xD;&#xA;del Y*&#xD;&#xA;del Z*&#xD;&#xA;del Eco.Core.dll&#xD;&#xA;del Eco.Gameplay.dll&#xD;&#xA;del Eco.Modkit.dll&#xD;&#xA;del Eco.Mods.dll&#xD;&#xA;del Eco.Networking.ENet.dll&#xD;&#xA;del Eco.Plugins.dll&#xD;&#xA;del Eco.Shared.dll&#xD;&#xA;del Eco.Simulation.dll&#xD;&#xA;del Eco.Stats.dll&#xD;&#xA;del Eco.Tests.dll&#xD;&#xA;del Eco.WebServer.dll&#xD;&#xA;del Eco.WindowsForms.dll&#xD;&#xA;del Eco.World.dll&#xD;&#xA;del Eco.WorldGenerator.dll" />
+	</Target>
 
-  <Target Name="PostBuild" AfterTargets="PostBuildEvent" Condition="'$(OS)' == 'Windows_NT'">
-    <Exec Command="cd $(TargetDir)&#xD;&#xA;del *.config&#xD;&#xA;del *.json&#xD;&#xA;del A*&#xD;&#xA;del B*&#xD;&#xA;del C*&#xD;&#xA;del D*&#xD;&#xA;del F*&#xD;&#xA;del G*&#xD;&#xA;del H*&#xD;&#xA;del I*&#xD;&#xA;del J*&#xD;&#xA;del K*&#xD;&#xA;del L*&#xD;&#xA;del M*&#xD;&#xA;del N*&#xD;&#xA;del O*&#xD;&#xA;del P*&#xD;&#xA;del Q*&#xD;&#xA;del R*&#xD;&#xA;del S*&#xD;&#xA;del T*&#xD;&#xA;del U*&#xD;&#xA;del V*&#xD;&#xA;del W*&#xD;&#xA;del X*&#xD;&#xA;del Y*&#xD;&#xA;del Z*&#xD;&#xA;del Eco.Core.dll&#xD;&#xA;del Eco.Gameplay.dll&#xD;&#xA;del Eco.Modkit.dll&#xD;&#xA;del Eco.Mods.dll&#xD;&#xA;del Eco.Networking.ENet.dll&#xD;&#xA;del Eco.Plugins.dll&#xD;&#xA;del Eco.Shared.dll&#xD;&#xA;del Eco.Simulation.dll&#xD;&#xA;del Eco.Stats.dll&#xD;&#xA;del Eco.Tests.dll&#xD;&#xA;del Eco.WebServer.dll&#xD;&#xA;del Eco.WindowsForms.dll&#xD;&#xA;del Eco.World.dll&#xD;&#xA;del Eco.WorldGenerator.dll" />
-  </Target>
+	<ItemGroup>
+		<Compile Remove="ConfigControllers\**" />
+		<EmbeddedResource Remove="ConfigControllers\**" />
+		<None Remove="ConfigControllers\**" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <Compile Remove="ConfigControllers\**" />
-    <EmbeddedResource Remove="ConfigControllers\**" />
-    <None Remove="ConfigControllers\**" />
-  </ItemGroup>
+	<ItemGroup>
+	  <None Remove="SpecialItems\largelumberStockpile.txt" />
+	  <None Remove="SpecialItems\lumberStockpile.txt" />
+	  <None Remove="SpecialItems\pickaxe.txt" />
+	  <None Remove="SpecialItems\smallStockpile.txt" />
+	  <None Remove="SpecialItems\Stockpile.txt" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <EmbeddedResource Include="Ecopedia\EcopediaStrings.csv" />
-    <EmbeddedResource Include="Ecopedia\ModDocumentation.xml" />
-  </ItemGroup>
+	<ItemGroup>
+		<EmbeddedResource Include="Ecopedia\EcopediaStrings.csv" />
+		<EmbeddedResource Include="Ecopedia\ModDocumentation.xml" />
+		<EmbeddedResource Include="SpecialItems\pickaxe.txt" />
+		<EmbeddedResource Include="SpecialItems\largelumberStockpile.txt" />
+		<EmbeddedResource Include="SpecialItems\lumberStockpile.txt" />
+		<EmbeddedResource Include="SpecialItems\smallStockpile.txt" />
+		<EmbeddedResource Include="SpecialItems\Stockpile.txt" />
+	</ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="ConcurrentHashSet" Version="1.3.0" />
+		<PackageReference Include="Eco.ReferenceAssemblies" Version="0.9.7.7-beta-release-431" />
+		<PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+		<PackageReference Include="NLog" Version="4.7.13" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Eco.ReferenceAssemblies" Version="0.9.6.1-beta-release-337" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc" Version="2.2.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-    <PackageReference Include="NLog" Version="4.7.13" />
-  </ItemGroup>
+	<ItemGroup>
+		<Folder Include="DiscordHandling\Framework\" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="DiscordHandling\Framework\" />
-  </ItemGroup>
+	<ItemGroup>
+	  <None Include="..\EMICON-text.png">
+	    <Pack>True</Pack>
+	    <PackagePath>\</PackagePath>
+	  </None>
+	  <None Include="..\README.md">
+	    <Pack>True</Pack>
+	    <PackagePath>\</PackagePath>
+	  </None>
+	</ItemGroup>
 
 </Project>
