@@ -1,29 +1,31 @@
 ﻿using Eco.EM.Framework.Logging;
+using Eco.Shared.Properties;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Net;
+using System.Net.Http;
 using System.Text;
+using System.Threading.Tasks;
 using static Eco.EM.Framework.WebHook;
 
 namespace Eco.EM.Framework.Networking
 {
     public static class Network
     {
+        private static readonly HttpClient httpClient = new();
+        private readonly static string jsonMediaType = "application/json";
+
+
         public static string GetRequest(string URL)
         {
             try
             {
-                var request = WebRequest.Create(URL) as HttpWebRequest;
-                request.CachePolicy = new System.Net.Cache.RequestCachePolicy(System.Net.Cache.RequestCacheLevel.NoCacheNoStore);
-                var response = (HttpWebResponse)request.GetResponse();
-
-                var header = response.Headers;
-
-                var encoding = Encoding.ASCII;
-                using var reader = new StreamReader(response.GetResponseStream(), encoding);
-                return reader.ReadToEnd();
+                httpClient.BaseAddress = new Uri(URL);
+                var result = httpClient.GetStringAsync(httpClient.BaseAddress);
+                return result.Result;
             }
             catch (Exception e)
             {
@@ -31,11 +33,12 @@ namespace Eco.EM.Framework.Networking
             }
         }
 
-        public static string PostRequest(string URL, Dictionary<string, string> Parameters)
+        public static async Task<string> PostRequestAsync(string URL, Dictionary<string, string> Parameters)
         {
             try
             {
-                var Request = WebRequest.Create(URL) as HttpWebRequest;
+                httpClient.BaseAddress = new Uri(URL);
+
                 var PostData = string.Empty;
 
                 foreach (var param in Parameters)
@@ -47,21 +50,10 @@ namespace Eco.EM.Framework.Networking
 
                     PostData += $"{param.Key}={param.Value}";
                 }
-                var data = Encoding.ASCII.GetBytes(PostData);
 
-                Request.Method = "POST";
-                Request.ContentType = "application/x-www-form-urlencoded";
-                Request.ContentLength = data.Length;
-
-                using (var stream = Request.GetRequestStream())
-                {
-                    stream.Write(data, 0, data.Length);
-                }
-
-                var Response = Request.GetResponse() as HttpWebResponse;
-                var ResponseString = new StreamReader(Response.GetResponseStream()).ReadToEnd();
-
-                return ResponseString;
+                StringContent strContent = new(PostData, Encoding.UTF8, jsonMediaType);
+                HttpResponseMessage responseMessage = await httpClient.PostAsync(httpClient.BaseAddress, strContent).ConfigureAwait(false);
+                return await responseMessage.Content.ReadAsStringAsync();
             }
             catch (Exception e)
             {
@@ -69,27 +61,16 @@ namespace Eco.EM.Framework.Networking
             }
         }
 
-        public static string PostRequest(string URL, string PostData)
+        public static async Task<string> PostRequest(string URL, string PostData)
         {
             try
             {
-                var Request = WebRequest.Create(URL) as HttpWebRequest;
+                httpClient.BaseAddress = new Uri(URL);
                 LoggingUtils.Debug(PostData);
-                var data = Encoding.ASCII.GetBytes(PostData);
 
-                Request.Method = "POST";
-                Request.ContentType = "application/json";
-                Request.ContentLength = data.Length;
-
-                using (Stream stream = Request.GetRequestStream())
-                {
-                    stream.Write(data, 0, data.Length);
-                }
-
-                var Response = Request.GetResponse() as HttpWebResponse;
-                var ResponseString = new StreamReader(Response.GetResponseStream()).ReadToEnd();
-
-                return ResponseString;
+                StringContent strContent = new StringContent(PostData, Encoding.UTF8, jsonMediaType);
+                HttpResponseMessage responseMessage = await httpClient.PostAsync(httpClient.BaseAddress, strContent).ConfigureAwait(false);
+                return await responseMessage.Content.ReadAsStringAsync();
             }
             catch (Exception e)
             {
