@@ -14,7 +14,7 @@ namespace Eco.EM.Framework.Permissions
     // Custom chat command processor assists in overriding SLG defined Auth levels and allows us to assign standard command to our own processing logic
     public class EMCustomChatProcessor : ICommandProcessorHandler
     {
-        public static Action<ChatCommand, IChatClient, bool> CommandProcessed;
+        private static readonly Func<ChatCommand, IChatClient, bool> commandProcessor;
 
         public EMCustomChatProcessor() { }
 
@@ -23,6 +23,11 @@ namespace Eco.EM.Framework.Permissions
         {
             var level = chatClient.GetChatAuthLevel();
             var adapter = CommandGroupsManager.FindAdapter(command.Name);
+            if (adapter == null)
+            {
+                chatClient.ErrorLocStr(string.Format(Defaults.appName + Localizer.DoStr("Command {0} not found"), command.Name));
+                return false;
+            }
 
             // if an admin or developer & we have not overridden this in our config return true;
             if ((level >= ChatAuthorizationLevel.Admin && CommandGroupsManager.Config.DefaultAdminBehaviour) || (command.AuthLevel == ChatAuthorizationLevel.User && CommandGroupsManager.Config.DefaultUserBehaviour))
@@ -30,28 +35,22 @@ namespace Eco.EM.Framework.Permissions
                 // Check For Blacklisted commands
                 if (GroupsManager.API.CommandPermitted(chatClient, adapter))
                 {
-                    CommandProcessed?.Invoke(command, chatClient, true);
+                    commandProcessor?.Invoke(command, chatClient);
                     return true;
                 }
-            }
-
-            if (adapter == null)
-            {
-                chatClient.ErrorLocStr(string.Format(Defaults.appName + Localizer.DoStr("Command {0} not found"), command.Name));
-                return false;
             }
 
             // check the users groups permissions permissions
             if (GroupsManager.API.UserPermitted(chatClient, adapter))
             {
-                CommandProcessed?.Invoke(command, chatClient, true);
+                commandProcessor?.Invoke(command, chatClient);
                 return true;
             }
 
             // default behaviour is to deny if command or user is not set
             chatClient.ErrorLocStr(string.Format(Defaults.appName + Localizer.DoStr("You are not authorized to use the command {0}"), command.Name));
 
-            CommandProcessed?.Invoke(command, chatClient, false);
+            commandProcessor?.Invoke(command, chatClient);
             return false;
         }
     }
