@@ -9,13 +9,13 @@ namespace Eco.EM.Framework.Resolvers
 {
     public class HousingModel : ModelBase
     {
-        [LocDisplayName("Room Type")]                     public HomeFurnishingValue.RoomCategory RoomType          { get; set; }
+        [LocDisplayName("Room Type")]                     public string RoomType          { get; set; }
         [LocDisplayName("Skill Value")]                   public float SkillValue                                   { get; set; }
         [LocDisplayName("Item Type")]                     public string TypeForRoomLimit                            { get; set; }
         [LocDisplayName("Diminishing Percent")]           public float DiminishingReturn                            { get; set; }
         [LocDisplayName("Item Name"), ReadOnly(true)]     public string DisplayName                                 { get; set; }
 
-        public HousingModel(Type type, string displayName, HomeFurnishingValue.RoomCategory roomType, float skillValue, string typeForRoomLimit, float diminishingReturn)
+        public HousingModel(Type type, string displayName, string roomType, float skillValue, string typeForRoomLimit, float diminishingReturn)
         {
             ModelType = type.Name;
             Assembly = type.AssemblyQualifiedName;
@@ -27,7 +27,7 @@ namespace Eco.EM.Framework.Resolvers
         }
 
         [JsonConstructor]
-        public HousingModel(string modelType, string assembly, HomeFurnishingValue.RoomCategory roomType, int skillValue, string typeForRoomLimit, float diminishingReturn, string displayName)
+        public HousingModel(string modelType, string assembly, string roomType, int skillValue, string typeForRoomLimit, float diminishingReturn, string displayName)
         {
             ModelType = modelType;
             Assembly = assembly;
