@@ -1,4 +1,5 @@
 ﻿using Eco.EM.Framework.Utils;
+using Eco.Mods.TechTree;
 using Eco.Shared.Localization;
 using Newtonsoft.Json;
 using System;
@@ -9,6 +10,7 @@ namespace Eco.EM.Framework.Resolvers
     public class StorageSlotModel : ModelBase
     {
         [LocDisplayName("Storage Slots")] public float StorageSlots { get; set; }
+        [LocDisplayName("Storage Stack Multiplier")] public float StackMultiplier {get; set;}
 
         public StorageSlotModel(Type type)
         {
@@ -17,11 +19,12 @@ namespace Eco.EM.Framework.Resolvers
         }
 
         [JsonConstructor]
-        public StorageSlotModel(string modelType, string assembly, float storageSlots)
+        public StorageSlotModel(string modelType, string assembly, float storageSlots, float stackMultiplier)
         {
             ModelType = modelType;
             Assembly = assembly;
             StorageSlots = storageSlots;
+            StackMultiplier = stackMultiplier;
         }
 
         public override string ToString() => $"{StringUtils.GetAssemblyNameFromAssemblyString(Assembly)} - {ModelType}";
