@@ -258,8 +258,8 @@ namespace Eco.EM.Framework.Utils
 
         public AffineTransform RotateX(float theta)
         {
-            float cot = Mathf.Cos(theta);
-            float sit = Mathf.Sin(theta);
+            float cot = MathF.Cos(theta);
+            float sit = MathF.Sin(theta);
             return Concatenate(
                     new AffineTransform(
                             1, 0, 0, 0,
@@ -275,8 +275,8 @@ namespace Eco.EM.Framework.Utils
 
         public AffineTransform RotateY(float theta)
         {
-            float cot = Mathf.Cos(theta);
-            float sit = Mathf.Sin(theta);
+            float cot = MathF.Cos(theta);
+            float sit = MathF.Sin(theta);
             return Concatenate(
                     new AffineTransform(
                             cot, 0, sit, 0,
@@ -292,8 +292,8 @@ namespace Eco.EM.Framework.Utils
 
         public AffineTransform RotateZ(float theta)
         {
-            float cot = Mathf.Cos(theta);
-            float sit = Mathf.Sin(theta);
+            float cot = MathF.Cos(theta);
+            float sit = MathF.Sin(theta);
             return Concatenate(
                     new AffineTransform(
                             cot, -sit, 0, 0,
